requirements.txt
flask
scikit-learn
pandas
joblib
pytesseract
Pillow
newspaper3k
